+++ ScanLineRender/ScanLineRender/AEL.h
@@ -26,14 +26,14 @@ typedef struct _EdgeListEntry {
 
 #define INIT_ELN_EZ(n, ev, ov) INIT_ELN(n, ev, ov, false)
 
-int16_t getMinXForLine(const EdgeListEntry * node, const int16_t scanLine);
-int16_t getMaxXForLine(const EdgeListEntry * node, const int16_t scanLine);
-int16_t getSmartXForLine(const EdgeListEntry * node, const int16_t scanLine);
+int32_t getMinXForLine(const EdgeListEntry * node, const int32_t scanLine);
+int32_t getMaxXForLine(const EdgeListEntry * node, const int32_t scanLine);
+int32_t getSmartXForLine(const EdgeListEntry * node, const int32_t scanLine);
 bool nodeHoldsSingleton(const EdgeListEntry * node);
 
 typedef struct {
 	LinkN* activeEdges;
-	int16_t scanLine;
+	int32_t scanLine;
 } ActiveEdgeList;
 
 const ActiveEdgeList freshAEL(void);
