+++ ScanLineRender/ScanLineRender/Point.h
@@ -16,7 +16,7 @@
 #define max(a, b) (((a) > (b)) ? (a) : (b))
 
 typedef struct {
-	int16_t x; int16_t y; int16_t z;
+	int32_t x; int32_t y; int32_t z;
 } Point;
 #define INIT_POINT(p, xv, yv, zv) \
 (p).x = (xv); \
